+++ src/hooks/automation/useAutomateButtons.js
@@ -1,11 +1,10 @@
-import { addToast } from '@heroui/react';
 import { invoke } from '@tauri-apps/api/core';
 import { useContext } from 'react';
 
 import { StateContext } from '@/components/contexts/StateContext';
 import { UserContext } from '@/components/contexts/UserContext';
-import ErrorToast from '@/components/ui/ErrorToast';
-import { logEvent } from '@/utils/global/tasks';
+import { checkSteamStatus, logEvent } from '@/utils/global/tasks';
+import { showDangerToast, showEnableAllGamesToast, showMissingCredentialsToast, showNoGamesToast, showOutdatedCredentialsToast } from '@/utils/global/toasts';
 
 // Automate card farming and achievement unlocking
 export const useAutomate = () => {
@@ -14,31 +13,16 @@ export const useAutomate = () => {
     // Start card farming
     const startCardFarming = async () => {
         try {
-            // Retrieve user summary from local storage
             // Check if Steam is running
-            const steamRunning = await invoke('is_steam_running');
+            const isSteamRunning = checkSteamStatus(true);
+            if (!isSteamRunning) return;
             // Retrieve Steam cookies from local storage
             const steamCookies = JSON.parse(localStorage.getItem('steamCookies')) || {};
             // Retrieve settings from local storage
             const settings = JSON.parse(localStorage.getItem('settings')) || {};
-            if (!steamRunning) {
-                return addToast({
-                    description: <ErrorToast
-                        message='Steam is not running'
-                        href='https://steamgameidler.vercel.app/faq#error-messages:~:text=Steam%20is%20not%20running'
-                    />,
-                    color: 'danger'
-                });
-            }
-            if (!steamCookies?.sid || !steamCookies?.sls) {
-                return addToast({
-                    description: <ErrorToast
-                        message='Missing credentials in Settings'
-                        href='https://steamgameidler.vercel.app/faq#error-messages:~:text=%22Missing%20credentials%20in%20setting%22'
-                    />,
-                    color: 'danger'
-                });
-            }
+
+            if (!steamCookies?.sid || !steamCookies?.sls) return showMissingCredentialsToast();
+
             // Validate Steam session
             const res = await invoke('validate_session', {
                 sid: steamCookies?.sid, sls: steamCookies?.sls, sma: steamCookies?.sma, steamid: userSummary.steamId
@@ -46,28 +30,15 @@ export const useAutomate = () => {
             if (!res.user) {
                 localStorage.removeItem('steamCookies');
                 localStorage.removeItem('cardFarmingUser');
-                return addToast({
-                    description: <ErrorToast
-                        message='Steam credentials need to be updated'
-                        href='https://steamgameidler.vercel.app/faq#error-messages:~:text=%22Steam%20credentials%20need%20to%20be%20updated%22'
-                    />,
-                    color: 'danger'
-                });
+                return showOutdatedCredentialsToast();
             }
             // Retrieve card farming list
             const cardFarmingList = await invoke('get_custom_lists', { steamId: userSummary.steamId, list: 'cardFarmingList' });
-            if (!settings.cardFarming.allGames && cardFarmingList.list_data.length === 0) {
-                return addToast({
-                    description: <ErrorToast
-                        message='Enable the "All games" setting or add some games to your card farming list'
-                        href='https://steamgameidler.vercel.app/faq#error-messages:~:text=Enable%20the%20%22All%20games%22%20setting%20or%20add%20some%20games%20to%20your%20card%20farming%20list'
-                    />,
-                    color: 'danger'
-                });
-            }
+            if (!settings.cardFarming.allGames && cardFarmingList.list_data.length === 0)
+                return showEnableAllGamesToast();
             setIsCardFarming(true);
         } catch (error) {
-            addToast({ description: `Error in (startCardFarming): ${error?.message || error}`, color: 'danger' });
+            showDangerToast('An error occurred. Check the logs for more information');
             console.error('Error in (startCardFarming):', error);
             logEvent(`[Error] in (startCardFarming): ${error}`);
         }
@@ -77,35 +48,24 @@ export const useAutomate = () => {
     const startAchievementUnlocker = async () => {
         try {
             // Check if Steam is running
-            const steamRunning = await invoke('is_steam_running');
+            const isSteamRunning = checkSteamStatus(true);
+            if (!isSteamRunning) return;
+
             // Retrieve settings from local storage
             const settings = JSON.parse(localStorage.getItem('settings')) || {};
-            if (!steamRunning) {
-                return addToast({
-                    description: <ErrorToast
-                        message='Steam is not running'
-                        href='https://steamgameidler.vercel.app/faq#error-messages:~:text=Steam%20is%20not%20running'
-                    />,
-                    color: 'danger'
-                });
-            }
+
             if (!settings || Object.keys(settings).length === 0) {
-                return addToast({ description: 'Please configure the settings first', color: 'danger' });
+                return showDangerToast('Please configure the settings first');
             }
+
             // Retrieve achievement unlocker list
             const achievementUnlockerList = await invoke('get_custom_lists', { steamId: userSummary.steamId, list: 'achievementUnlockerList' });
-            if (achievementUnlockerList.list_data.length === 0) {
-                return addToast({
-                    description: <ErrorToast
-                        message='There are no games in your achievement unlocker list'
-                        href='https://steamgameidler.vercel.app/faq#error-messages:~:text=Steam%20is%20not%20running'
-                    />,
-                    color: 'danger'
-                });
-            }
+            if (achievementUnlockerList.list_data.length === 0)
+                return showNoGamesToast();
+
             setIsAchievementUnlocker(true);
         } catch (error) {
-            addToast({ description: `Error in (startAchievementUnlocker): ${error?.message || error}`, color: 'danger' });
+            showDangerToast('An error occurred. Check the logs for more information');
             console.error('Error in (startAchievementUnlocker):', error);
             logEvent(`[Error] in (startAchievementUnlocker): ${error}`);
         }